+++ src/omniimage/__init__.py
@@ -1,4 +1,4 @@
 """OmniImage Converter package."""
 
 __app_name__ = "OmniImage Converter"
-__version__ = "0.3.0"
+__version__ = "0.3.1"